requirements.txt
scandir
conda
imglib2-imglyb
Pyjnius
